Server/requirements.txt
fastapi[standard]
nltk
numpy
pandas
pillow
scikit-learn
tokenizers
torch
torchvision
transformers
Werkzeug
ultralytics
openai
